Audit function fixed

+++ Crypto/mpi/part.h
@@ -13,9 +13,9 @@
 #include "mem.h"
 #include "struct.h"
 
-#define DISPOSITIONS(cs_len, pass_len) (powl((cs_len),(pass_len)));
+#define DISPOSITIONS(cs_len, pass_len) (pow((cs_len),(pass_len)));
 #define DISP_PER_PROC(disp, num_procs) (((disp)/(num_procs))+1);
-#define STARTING_CHAR(init, cs_size, pos) (((long)((init)/(powl(cs_size, pos))))%(cs_size));
+#define STARTING_CHAR(init, cs_size, pos) (((long)((init)/(pow(cs_size, pos))))%(cs_size));
 
 #define BCAST_TAG 0
